requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
_tflow_select=2.3.0=mkl
abseil-cpp=20210324.2=h2531618_0
absl-py=0.15.0=pyhd3eb1b0_0
aiohttp=3.8.1=py39h7f8727e_1
aiosignal=1.2.0=pyhd3eb1b0_0
appdirs=1.4.4=pyhd3eb1b0_0
astor=0.8.1=py39h06a4308_0
astunparse=1.6.3=py_0
async-timeout=4.0.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
black=19.10b0=py_0
blas=1.0=mkl
blinker=1.4=py39h06a4308_0
blosc=1.21.0=h8c45485_0
bottleneck=1.3.4=py39hce1f21e_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py39h27cfd23_1003
brunsli=0.1=h2531618_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f8727e_0
ca-certificates=2022.6.15=ha878542_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.6.15=py39hf3d152e_0
cffi=1.15.0=py39hd667e15_1
cfitsio=3.470=hf0d0db6_6
charls=2.2.0=h2531618_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cleverhans=4.0.0=pypi_0
click=8.0.4=py39h06a4308_0
cloudpickle=2.1.0=pypi_0
cryptography=3.4.8=py39hd23ed53_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
cytoolz=0.11.0=py39h27cfd23_0
dask-core=2022.7.0=py39h06a4308_0
dataclasses=0.8=pyh6d0b6a4_7
decorator=5.1.1=pypi_0
dm-tree=0.1.7=pypi_0
easydict=1.9=pypi_0
et_xmlfile=1.1.0=py39h06a4308_0
ffmpeg=4.3=hf484d3e_0
flatbuffers=2.0.0=h2531618_0
fonttools=4.33.3=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.2.0=py39h7f8727e_0
fsspec=2022.3.0=py39h06a4308_0
gast=0.5.3=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.42.0=py39hce63b2e_0
h5py=3.6.0=py39ha0f2276_0
hdf5=1.10.6=hb1b8bf9_0
icu=68.1=h2531618_0
idna=3.3=pyhd3eb1b0_0
imagecodecs=2021.8.26=py39h4cda21f_0
imageio=2.19.3=py39h06a4308_0
importlib-metadata=4.11.3=py39h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h7f8727e_0
jxrlib=1.1=h7b6447c_2
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.4.2=pypi_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libcurl=7.82.0=h0b77cf5_0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libzopfli=1.0.3=he6710b0_0
locket=1.0.0=py39h06a4308_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py39h06a4308_0
matplotlib=3.5.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mnist=0.2.2=pypi_0
monai=0.9.0=pypi_0
monai-weekly=0.10.dev2227=pypi_0
multidict=5.2.0=py39h7f8727e_2
mypy_extensions=0.4.3=py39h06a4308_1
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py39h06a4308_0
nose=1.3.7=pypi_0
numexpr=2.8.1=py39h6abb31d_0
numpy=1.21.5=py39he7a7128_1
numpy-base=1.21.5=py39hf524024_1
oauthlib=3.2.0=pyhd3eb1b0_0
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openpyxl=3.0.9=pyhd3eb1b0_0
openssl=1.1.1q=h7f8727e_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py39h295c915_0
partd=1.2.0=pyhd3eb1b0_1
pathspec=0.7.0=py_0
pillow=9.0.1=py39h22f2fdc_0
pip=21.2.4=py39h06a4308_0
protobuf=3.14.0=py39h2531618_1
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycodestyle=2.8.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydicom=2.3.0=pyh6c4a22f_0
pyjwt=2.1.0=py39h06a4308_0
pyopenssl=21.0.0=pyhd3eb1b0_1
pyparsing=3.0.4=pyhd3eb1b0_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.12=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-flatbuffers=1.12=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.3.0=py39h7f8727e_0
pyyaml=6.0=py39h7f8727e_1
readline=8.1.2=h7f8727e_1
regex=2022.3.15=py39h7f8727e_0
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-image=0.19.2=py39h51133e4_0
scikit-learn=1.0.2=py39h51133e4_1
scipy=1.7.3=py39hc147768_0
setuptools=61.2.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlite=3.38.2=hc218d9a_0
tensorboard=2.6.0=py_1
tensorboard-data-server=0.6.0=py39hca6d32c_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.6.0=mkl_py39haac40d1_0
tensorflow-base=2.6.0=mkl_py39h3d85931_0
tensorflow-estimator=2.6.0=pyh7b7c402_0
tensorflow-probability=0.16.0=pypi_0
termcolor=1.1.0=py39h06a4308_1
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.11=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.11.2=pyhd3eb1b0_0
torchaudio=0.11.0=py39_cu102
torchvision=0.12.0=py39_cu102
tqdm=4.64.0=py39h06a4308_0
typed-ast=1.4.3=py39h7f8727e_1
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39h06a4308_0
werkzeug=2.0.3=pyhd3eb1b0_0
wget=3.2=pypi_0
wheel=0.35.1=pyhd3eb1b0_0
wrapt=1.13.3=py39h7f8727e_2
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py39h27cfd23_0
zfp=0.5.5=h295c915_6
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=haebb681_0
